+++ bower.json
@@ -1,23 +1,20 @@
 {
   "name": "ember-twiddle",
   "dependencies": {
+    "ember": "1.13.3",
+    "ember-cli-shims": "ember-cli/ember-cli-shims#0.0.3",
+    "ember-cli-test-loader": "ember-cli-test-loader#0.1.3",
+    "ember-data": "1.13.5",
+    "ember-load-initializers": "ember-cli/ember-load-initializers#0.1.5",
+    "ember-qunit": "0.4.1",
+    "ember-qunit-notifications": "0.0.7",
+    "ember-resolver": "~0.1.18",
+    "jquery": "^1.11.1",
+    "loader.js": "ember-cli/loader.js#3.2.0",
+    "qunit": "~1.17.1",
     "codemirror": "~5.3.0",
     "ember-cli-codemirror-shim": "~0.0.1",
     "bootstrap-sass-official": "~3.3.4",
-    "dom-ruler": "0.1.5",
-    "ember": "1.13.0",
-    "ember-cli-shims": "ember-cli/ember-cli-shims#0.0.3",
-    "ember-cli-test-loader": "ember-cli-test-loader#0.1.3",
-    "ember-data": "1.0.0-beta.18",
-    "ember-load-initializers": "ember-cli/ember-load-initializers#0.1.4",
-    "ember-qunit": "0.3.3",
-    "ember-qunit-notifications": "0.0.7",
-    "ember-resolver": "~0.1.17",
-    "jquery": "^1.11.1",
-    "loader.js": "ember-cli/loader.js#3.2.0",
-    "qunit": "~1.17.1"
-  },
-  "resolutions": {
-    "codemirror": "~5.3.0"
+    "dom-ruler": "0.1.5"
   }
 }
