+++ package.json
@@ -18,37 +18,36 @@
   },
   "author": "",
   "license": "MIT",
-
-  "dependencies": {
-    "babel": "^5.0.0",
+  "devDependencies": {
     "broccoli-asset-rev": "^2.0.2",
+    "ember-cli": "1.13.1",
+    "ember-cli-app-version": "0.4.0",
+    "ember-cli-babel": "^5.0.0",
+    "ember-cli-dependency-checker": "^1.0.0",
+    "ember-cli-htmlbars": "0.7.9",
+    "ember-cli-htmlbars-inline-precompile": "^0.1.1",
+    "ember-cli-ic-ajax": "0.2.1",
+    "ember-cli-inject-live-reload": "^1.3.0",
+    "ember-cli-qunit": "0.3.15",
+    "ember-cli-release": "0.2.3",
+    "ember-cli-uglify": "^1.0.1",
+    "ember-data": "1.13.5",
+    "ember-disable-proxy-controllers": "^1.0.0",
+    "ember-export-application-global": "^1.0.2",
+    "babel": "^5.0.0",
+    "ember-browserify": "^1.0.0",
+    "ember-cli-deploy": "^0.4.2",
+    "ember-deploy-s3": "0.0.4",
+    "ember-deploy-s3-index": "^0.4.0",
+    "ember-cli-bootstrap-sassy": "0.0.12",
+    "ember-cli-build-notifications": "0.0.3",
+    "ember-cli-file-creator": "^0.1.0",
     "broccoli-sass": "^0.6.6",
     "dc-tabs": "0.1.0",
     "dedent": "^0.4.0",
     "ember-autoresize": "0.4.1",
-    "ember-browserify": "^1.0.0",
-    "ember-cli": "0.2.7",
-    "ember-cli-app-version": "0.3.3",
-    "ember-cli-babel": "^5.0.0",
-    "ember-cli-bootstrap-sassy": "0.0.12",
-    "ember-cli-build-notifications": "0.0.3",
-    "ember-cli-file-creator": "^0.1.0",
-    "ember-cli-dependency-checker": "^1.0.0",
-    "ember-cli-htmlbars": "0.7.9",
-    "ember-cli-ic-ajax": "0.1.1",
-    "ember-cli-inject-live-reload": "^1.3.0",
-    "ember-cli-qunit": "0.3.13",
-    "ember-cli-release": "0.2.3",
-    "ember-cli-uglify": "^1.0.1",
-    "ember-data": "1.0.0-beta.18",
     "ember-notify": "3.3.2",
     "ivy-codemirror": "^1.1.0",
-    "liquid-fire": "^0.20.0",
     "torii": "^0.3.5"
-  },
-  "devDependencies": {
-    "ember-cli-deploy": "^0.4.2",
-    "ember-deploy-s3": "0.0.4",
-    "ember-deploy-s3-index": "^0.4.0"
   }
 }
